+++ Types/UIImageBasedButton.cs
@@ -1,11 +1,12 @@
-﻿using UnityEngine;
-using UnityEngine.Events;
+﻿using System;
+using UnityEngine;
 using UnityEngine.EventSystems;
 using UnityEngine.UI;
 
 namespace MyBox
 {
-	public class UIImageBasedButton : MonoBehaviour, ISelectHandler, IDeselectHandler
+	[RequireComponent(typeof(Button), typeof(Image))]
+	public class UIImageBasedButton : MonoBehaviour, ISelectHandler, IDeselectHandler, IPointerEnterHandler, IPointerExitHandler
 	{
 #pragma warning disable 0649
 		[SerializeField, MustBeAssigned] private Sprite _regularSprite;
@@ -14,40 +15,34 @@ namespace MyBox
 		[SerializeField, MustBeAssigned] private Sprite _clickedSelectedSprite;
 #pragma warning restore 0649
 
-
+		public Action<bool> OnToggled;
+		
 		public bool AlternativeSpriteset
 		{
-			get { return _alternative; }
-			set { _alternative = value; }
+			get => _alternative;
+			set => _alternative = value;
 		}
 
 		private bool _alternative;
 		private bool _selected;
 		private Image _image;
 		private Button _button;
-		private UnityAction _event;
 
 
 		private void Awake()
 		{
 			_image = GetComponent<Image>();
 			_button = GetComponent<Button>();
-			_event = ToggleSprites;
 		}
 
-		private void OnEnable()
-		{
-			_button.onClick.AddListener(_event);
-		}
-
-		private void OnDisable()
-		{
-			_button.onClick.RemoveListener(_event);
-		}
+		private void OnEnable() => _button.onClick.AddListener(ToggleSprites);
+		private void OnDisable() => _button.onClick.RemoveListener(ToggleSprites);
 
 		private void ToggleSprites()
 		{
 			_alternative = !_alternative;
+			OnToggled?.Invoke(_alternative);
+			
 			UpdateSprites();
 		}
 
@@ -71,6 +66,16 @@ namespace MyBox
 		}
 
 		public void OnDeselect(BaseEventData eventData)
+		{
+			UpdateSprites(false);
+		}
+
+		public void OnPointerEnter(PointerEventData eventData)
+		{
+			UpdateSprites(true);
+		}
+
+		public void OnPointerExit(PointerEventData eventData)
 		{
 			UpdateSprites(false);
 		}