+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this package will be documented in this file.
 - Breaking Changes: MyCollections.GetOrDefault renamed to GetOrAdd
 - Breaking Changes: MyCollections.SelectWithIndex and SelectManyWithIndex removed (Linq.Select and SelectMany functionality duplication)
 - Breaking Changes: MonoSingleton class is removed, its functionality combined with Singleton class
+- Breaking Changes: UIImageBasedToggle removed. Unity's Toggle able to do same things this days
 - Added: Donation button in MyBox window ;)
 - Added: TransformData type to store and restore position, rotation and scale
 - Added: CharactersRange Attribute to validate string by set of characters
@@ -21,6 +22,7 @@ All notable changes to this package will be documented in this file.
 - Changed: AutoPropertyAttribute will also check prefabs on prefab mode open
 - Changed: MyBox will only check for updates when editor opens
 - Changed: Singleton now might be used as parent class of MonoBehaviour to cache and remove duplicating instances
+- Changed: UIImageBasedButton now uses Highlight sprites when Pointed
 - Fix: DisplayInspectorAttribute works better with (and inside of) collections
 - Fix: MyString.ToCamelCase() extension works better now, thanks to @derfium!
 - Fix: MinValue/MaxValueAttribute build warnings fixed